scripts/requirements.txt
psycopg2-binary==2.9.9
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0
Pillow==10.2.0
